Allow continuation after Eof has been reached
The previous implementation of quick-xml considered that when the
underlying `BufRead` implementation returned no data, it meant that
the absolute end of the XML document end was reached.

This is problematic for us (as developpers of the RxPlayer media player)
as we may have temporary "BufRead starvation" in our use-cases.

---

This should be a very rare situation when considering only the Rust
language, but it can be a much more frequent one when using this library
conjointly with JavaScript, e.g. by compiling down this library to
WebAssembly.

---

Take our case: we want to be able to perform a request for an XML
document on the JavaScript-side, while processing the response through
this library at the same time (i.e. at the same time than bytes from
the request are received).

This means that we won't immediately have the full XML available when
calling into quick-xml, and as we're using JavaScript to obtain the
original data, we cannot just block the `BufRead` object given.

We thought of three potential solutions:

  1. We do not change anything in quick-xml and add a "security
     padding" big enough so that quick-xml will not be able to
     encounter the end of the document until the true end is available.

     This could be possible because quick-xml acts "event" per "event".
     We could judge that not enough data is available for the next event
     and thus not call quick-xml.

     However this solution is risky (what if our padding is not enough?)
     and could be inneficient in cases of large paddings.

  2. We could define a Future-based reader and use it instead in
     quick-xml. Here the Future would be "awaken" only when new data is
     available (and could be backed with a JavaScript-side Promise).

     This solution could however be much more complicated to implement.

  3. We update quick-xml so it can pick-up where it was even when an Eof
     was last encountered. This looks efficient and non-risky but it
     imply a much more torough understanding of quick-xml inner workings.

This commit is the implementation of the third one.

I try to test mainly my use-case but it is very possible that some other
methods don't behave like you would think in some edge scenarios (all
tests are passing though).

The biggest "ugly" thing left is that unlike before the `buf` argument
given to some of quick-xml methods cannot be `clear`ed at any time, for
reasons that are easy to guess (we need to conserve some more state than
before). For now, I added a `try_clear_buffer` method to still allow
clearing that buffer in most cases without breaking anything.